--- src/mock/event-offer.js
@@ -1,12 +1,0 @@
-import {tripOffersMap, TRANSFER_TYPE, ACTIVITY_TYPE} from '../utils/const';
-import {getRandomArrayItem} from './utils';
-
-export const generateEventOffers = () => {
-  const pointTypes = [...TRANSFER_TYPE, ...ACTIVITY_TYPE];
-  const type = getRandomArrayItem(pointTypes);
-
-  return {
-    type,
-    offers: tripOffersMap.get(type),
-  };
-};
